+++ test/util/generation.test.ts
@@ -1,6 +1,6 @@
-import { retryWithDeduplication, sampleArray } from '../src/util/generation';
+import { retryWithDeduplication, sampleArray } from '../../src/util/generation';
 
-jest.mock('../src/logger');
+jest.mock('../../src/logger');
 
 describe('retryWithDeduplication', () => {
   it('should collect unique items until target count is reached', async () => {
